Squashed commit of the following:
commit 51f70727ae4017ec363d03b02266b20c7a1fb9e4
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Mon Mar 30 00:34:31 2026 +0900

    refactor: 로그 숨김

commit fcb0c8f773ebe829ef57c40dfcb26a557b10579d
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Sun Mar 29 16:00:39 2026 +0900

    refactor

commit 1a890f1378b900b623e3aca4fdfc42f7e6af19f1
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Wed Mar 25 16:58:57 2026 +0900

    feat: 유저 레디 적용

commit 02612cd111ca570021aa57b82b7d9612fb1899eb
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Wed Mar 25 00:57:35 2026 +0900

    feat: Transform과 MovementComponent 추가

commit 9aa49a667c76065a9e53d27d834f842e17ea2fd4
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Tue Mar 24 17:55:06 2026 +0900

    refactor: 코드 정리

commit 238cbdf750e55f1fae0ee7466a99b4be07af1211
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Sun Mar 22 19:42:28 2026 +0900

    refactor: 코드 정리

commit badb19af0f06a894fbb28aeee724d1d856e4345f
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Tue Mar 17 20:32:44 2026 +0900

    refactor: Legacy 코드 제거

commit d790058190bfabe2739622954f2e51a041cb806e
Author: Seongyeol Yun <syyundev@gmail.com>
Date:   Tue Mar 17 19:41:28 2026 +0900

    refactor: ServerEngine Legacy 코드 제거

+++ EisenValor-Client/EisenValor-ClientFramework/src/SceneGlobal.cpp
@@ -53,7 +53,7 @@ Scene* SceneGlobal::GetScene(const std::string& sceneName) const
 	return nullptr;
 }
 
-void SceneGlobal::SetLocalNetworkID(uint32_t networkID) 
+void SceneGlobal::SetLocalNetworkID(uint64_t networkID) 
 {
 	m_localNetworkID = networkID;
